Saving progress on SudokuBoard refactoring
