+++ sudokuru/app/Components/Sudoku Board/SudokuBoard.tsx
@@ -14,6 +14,15 @@ import {useFocusEffect} from "@react-navigation/core";
 import {PreferencesContext} from "../../Contexts/PreferencesContext";
 import {useTheme} from "react-native-paper";
 import NumberControl from "./Components/NumberControl";
+import {
+  checkSolution,
+  formatTime, getBoxIndexFromXY, getCausesFromHint,
+  getCellNumber,
+  getCellSize, getDrillSolutionCells,
+  getGroupsFromHint,
+  getNumberOfGroupsAssignedForNumber, getPlacementsFromHint, getRemovalsFromHint, parseApiAndAddNotes, replaceChar,
+  strPuzzleToArray, updateBoardWithNumber, getHint
+} from "./BoardFunctions";
 
 // Sudokuru Package Import
 const sudokuru = require("../../../node_modules/sudokuru/dist/bundle.js");
@@ -195,144 +204,6 @@ const styles = (cellSize, sizeConst, theme) => StyleSheet.create({
   }
 });
 
-// USAGE
-// board = addNumberAsNote(...)
-function addNumberAsNote (number, board, i, j) {
-  let selectedCell = board.get('puzzle').getIn([i, j]);
-  if (!selectedCell)
-  {
-    return;
-  }
-  const prefilled = selectedCell.get('prefilled');
-  if (prefilled)
-  {
-    return;
-  }
-  let notes = selectedCell.get('notes') || new Set();
-  notes = notes.add(number);
-  selectedCell = selectedCell.set('notes', notes);
-  board = board.setIn(['puzzle', i, j], selectedCell);
-  return board;
-};
-
-function componentBoardValsToArray(board)
-{
-  let boardArray = [];
-  let temp = [];
-  for (let i = 0; i < 9; i++)
-  {
-    temp = [];
-    for (let j = 0; j < 9; j++)
-    {
-      currVal = board.get('puzzle').getIn([i, j, 'value']);
-      temp.push(!currVal ? "0" : currVal.toString());
-    }
-    boardArray.push(temp);
-  }
-  return boardArray;
-}
-
-function componentBoardNotesToArray(board)
-{
-  let notesArray = [];
-  let temp = [];
-  for (let i = 0; i < 9; i++)
-  {
-    for (let j = 0; j < 9; j++)
-    {
-      temp = [];
-      let notesSetFromComponent = board.get('puzzle').getIn([i, j, 'notes']);
-      if (!notesSetFromComponent)
-      {
-        notesArray.push(temp);
-        continue;
-      }
-      for (let k = 1; k <= 9; k++)
-      {
-        if (notesSetFromComponent.includes(k))
-        {
-          temp.push((k).toString());
-        }
-      }
-      notesArray.push(temp);
-    }
-  }
-  return notesArray;
-}
-
-function componentSolutionValsToArray(solution)
-{
-  let solArray = [];
-  let temp = [];
-  for (let i = 0; i < 9; i++)
-  {
-    temp = [];
-    for (let j = 0; j < 9; j++)
-    {
-      temp.push(solution[9 * j + i].toString());
-    }
-    solArray.push(temp);
-  }
-  return solArray;
-}
-
-function getHint(board, solution, strategies)
-{
-  let hintStrategies = [...strategies];
-  hintStrategies.push("AMEND_NOTES", "SIMPLIFY_NOTES");
-  let boardArray = componentBoardValsToArray(board);
-  let notesArray = componentBoardNotesToArray(board);
-  let solutionArray = componentSolutionValsToArray(solution);
-  let hint;
-  try {
-    hint = Puzzles.getHint(boardArray, notesArray, hintStrategies, solutionArray);
-  } catch (e) {
-    console.log(e);
-  }
-  return hint;
-}
-
-function parseApiAndAddNotes(board, puzzleCurrentNotesState, isDrill)
-{
-  if (!puzzleCurrentNotesState)
-  {
-    return;
-  }
-  if (puzzleCurrentNotesState.length != 729)
-  {
-    return;
-  }
-  let stringIndex = 0;
-  for (let i = 0; i < 9; i++)
-  {
-    for (let j = 0; j < 9; j++)
-    {
-      for (let currNoteIndex = 0; currNoteIndex < 9; currNoteIndex++)
-      {
-        stringIndex = 81 * i + 9 * j + currNoteIndex;
-
-        if (puzzleCurrentNotesState.charAt(stringIndex) == 1){
-          if (isDrill){
-            board = addNumberAsNote(currNoteIndex + 1, board, j, i);
-          } else {
-            board = addNumberAsNote(currNoteIndex + 1, board, i, j);
-          }
-        }
-      }
-    }
-  }
-  return board;
-}
-
-function strPuzzleToArray(str) {
-  let arr = [];
-  for (let i = 0; i < str.length; i += 9) {
-    arr.push(str.slice(i, i + 9).split('').map(Number));
-  }
-  let output = arr[0].map((_, colIndex) => arr.map(row => row[colIndex]));
-  return { puzzle: output };
-}
-
 async function generateGame(url, props) {
 
   let token = null;
@@ -424,58 +295,6 @@ async function generateGame(url, props) {
   return gameData;
 }
 
-// for each cell that is a part of the hint, store the coordinates and the resulting state
-// if there is a notes field for the cell, the notes must match
-// if there is a value field for the cell, the value must match
-function getDrillSolutionCells(board, solution, strategies)
-{
-  let drillSolutionCells = [];
-  let hint = getHint(board, solution, strategies);
-  if (hint)
-  {
-    for (let i = 0; i < hint.removals.length; i++)
-    {
-      let temp = {};
-      let currRemoval = hint.removals[i];
-      temp.x = currRemoval[0];
-      temp.y = currRemoval[1];
-      temp.notes = board.get('puzzle').getIn([temp.x, temp.y]).get('notes');
-      for (let j = 2; j < currRemoval.length; j++)
-      {
-        temp.notes = temp.notes.delete(currRemoval[j]);
-      }
-      drillSolutionCells.push(temp);
-    }
-
-    if (hint.placements[0])
-    {
-      let temp = {}
-      temp.x = hint.placements[0][0];
-      temp.y = hint.placements[0][1];
-      temp.value = hint.placements[0][2];
-      drillSolutionCells.push(temp);
-    }
-  }
-  return drillSolutionCells;
-};
-
-// https://stackoverflow.com/questions/36098913/convert-seconds-to-days-hours-minutes-and-seconds
-const formatTime = (inputSeconds: number) => {
-  // Get minutes and remaining seconds
-  const days = Math.floor(inputSeconds / (3600*24));
-  const hours = Math.floor(inputSeconds % (3600*24) / 3600);
-  const minutes = Math.floor(inputSeconds % 3600 / 60);
-  const seconds = Math.floor(inputSeconds % 60);
-  // Pad with zeros if needed
-  const paddedDays = days > 0 ? (days < 10 ? "0" : "") + days + ":" : "";
-  const paddedHours = hours > 0 ? (hours < 10 ? "0" : "") + hours + ":" : (hours == 0 && days != 0) ? "00" : "";
-  const paddedMinutes = minutes > 0 ? (minutes < 10 ? "0" : "") + minutes + ":" : (minutes == 0 && hours != 0) ? "00" : "";
-  const paddedSeconds = seconds > 0 ? (seconds < 10 ? "0" : "") + seconds : (seconds == 0 && minutes != 0) ? "00" : "0";
-
-  // Return formatted string
-  return `${paddedDays}${paddedHours}${paddedMinutes}${paddedSeconds}`;
-};
-
 const Puzzle = (props) => {
   const { board, renderCell } = props;
   const cellSize = getCellSize();
@@ -503,93 +322,6 @@ Puzzle.propTypes = {
   onFirstStep: PropTypes.bool,
   onFinalStep: PropTypes.bool,
 };
-
-
-// function that converts x,y cell coords to a number
-const getCellNumber = (x, y) => {
-  return y + x * 9;
-};
-
-const getBoxIndexFromCellNum = (cellNum) => {
-  return Math.floor((cellNum % 9) / 3);
-}
-
-const getBoxIndexFromXY = (x,y) => {
-  return Math.floor(x / 3) * 3 + Math.floor(y / 3);
-}
-
-const getCausesFromHint = (hint) => {
-  let causes = []
-  for (let i = 0; i < hint.cause.length; i++)
-  {
-    causes.push(hint.cause[i])
-  }
-  return causes
-}
-
-const getGroupsFromHint = (hint) => {
-  let groups = []
-  let temp = {}
-  for (let i = 0; i < hint.groups.length; i++)
-  {
-    temp = {}
-    switch (hint.groups[i][0])
-    {
-      case 0: // column
-        temp.type = "col"
-        temp.index = hint.groups[i][1]
-        break;
-      case 1: // row
-        temp.type = "row"
-        temp.index = hint.groups[i][1]
-        break;
-      case 2: // box
-        temp.type = "box"
-        temp.index = hint.groups[i][1]
-        break;
-    }
-    groups.push(temp)
-  }
-  return groups
-}
-
-const getPlacementsFromHint = (hint) => {
-  let placements = []
-  let temp = {}
-  for (let i = 0; i < hint.placements.length; i++)
-  {
-    temp = {}
-    temp.position = []
-    temp.position.push(hint.placements[i][0])
-    temp.position.push(hint.placements[i][1])
-    temp.value = hint.placements[i][2]
-    placements.push(temp)
-  }
-  return placements
-}
-
-const getRemovalsFromHint = (board, hint) => {
-  let removals = []
-  let temp = {}
-  for (let i = 0; i < hint.removals.length; i++)
-  {
-    let x = hint.removals[i][0]
-    let y = hint.removals[i][1]
-    temp = {}
-    temp.position = []
-    temp.position.push(x)
-    temp.position.push(y)
-    temp.values = []
-    temp.values.push()
-    for (let j = 2; j < hint.removals[i].length; j++)
-      temp.values.push(hint.removals[i][j])
-
-    removals.push(temp)
-  }
-  return removals
-}
-
-// let demoHighlightInput = [[0,7, darkBrown], [1,5, darkBrown], [2,0], [3, 4, 6, gold]];
 
 async function saveGame(activeGame) {
     let token = null;
@@ -619,20 +351,6 @@ async function finishGame(activeGame, showResults) {
           showResults(res.score, res.solveTime, res.numHintsUsed, res.numWrongCellsPlayed, res.difficulty);
         }
     });
-}
-
-function replaceChar(origString, replaceChar, index) {
-  let firstPart = origString.substr(0, index);
-  let lastPart = origString.substr(index + 1);
-
-  return firstPart + replaceChar + lastPart;
-}
-
-const checkSolution = (solution, x, y, value) => {
-  let cellNum = getCellNumber(y, x); // Flipping x and y because of how the solution string is formatted
-  let solutionValue = solution.charAt(cellNum);
-  
-  return solutionValue == value;
 }
 
 let puzzleString = "";
@@ -1056,57 +774,12 @@ const HintSection = (props) => {
   );
 }
 
-/*
- * This function retrieves the user's device size and calculates the cell size
- * board has width and height dimensions of 1 x 1.44444
- */
-export function getCellSize()
-{
-  const size = useWindowDimensions();
-  
-  return Math.min(size.width * 1.44444, size.height) / 15;
-}
-
-function updateBoardWithNumber({ x, y, number, fill = true, board }) {
-
-  let cell = board.get('puzzle').getIn([x, y]);
-  cell = cell.delete('notes');
-  cell = fill ? cell.set('value', number) : cell.delete('value');
-
-  const increment = fill ? 1 : -1;
-  const rowPath = ['choices', 'rows', x, number];
-  const columnPath = ['choices', 'columns', y, number];
-  const squarePath = ['choices', 'squares', ((Math.floor(x / 3)) * 3) + Math.floor(y / 3), number];
-  
-  return board.setIn(rowPath, board.getIn(rowPath) + increment)
-    .setIn(columnPath, board.getIn(columnPath) + increment)
-    .setIn(squarePath, board.getIn(squarePath) + increment)
-    .setIn(['puzzle', x, y], cell);
-}
-
-function getNumberOfGroupsAssignedForNumber(number, groups) {
-  return groups.reduce((accumulator, row) =>
-    accumulator + (row.get(number) > 0 ? 1 : 0), 0);
-}
-
-
-
+//todo convert this class component into a functional component
 export default class SudokuBoard extends React.Component<SudokuBoardProps> {
   constructor(props) {
     super(props);
   };
   state = generateGame(USERACTIVEGAMESBFFURL, this.props);
-
-  componentDidMount = () => {
-    if ('serviceWorker' in navigator) {
-      navigator.serviceWorker.register('/service-worker.js').then((reg) => {
-        console.log('ServiceWorker scope: ', reg.scope);
-        console.log('service worker registration successful');
-      }).catch((err) => {
-        console.warn('service worker registration failed', err.message);
-      });
-    }
-  }
 
   getSelectedCell = () => {
     const { board } = this.state;
